refactor: 패키지를 기능별로 구분하도록 변경
