+++ server/src/main/java/wap/starlist/bookmark/domain/ParentType.java
@@ -1,4 +1,4 @@
-package wap.starlist.domain;
+package wap.starlist.bookmark.domain;
 
 public enum ParentType {
     ROOT,
