+++ server/src/main/java/wap/starlist/bookmark/dto/response/BookmarkErrorResponse.java
@@ -1,4 +1,4 @@
-package wap.starlist.dto.response;
+package wap.starlist.bookmark.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
