+++ server/src/main/java/wap/starlist/bookmark/dto/response/BookmarkResponse.java
@@ -1,8 +1,8 @@
-package wap.starlist.dto.response;
+package wap.starlist.bookmark.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
-import wap.starlist.domain.Bookmark;
+import wap.starlist.bookmark.domain.Bookmark;
 
 @Getter
 @Builder
