+++ server/src/main/java/wap/starlist/bookmark/repository/BookmarkRepository.java
@@ -1,8 +1,8 @@
-package wap.starlist.repository;
+package wap.starlist.bookmark.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import wap.starlist.domain.Bookmark;
+import wap.starlist.bookmark.domain.Bookmark;
 
 @Repository
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
