+++ server/src/main/java/wap/starlist/bookmark/repository/FolderRepository.java
@@ -1,8 +1,8 @@
-package wap.starlist.repository;
+package wap.starlist.bookmark.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import wap.starlist.domain.Folder;
+import wap.starlist.bookmark.domain.Folder;
 
 @Repository
 public interface FolderRepository extends JpaRepository<Folder, Long> {
