+++ server/src/main/java/wap/starlist/bookmark/repository/RootRepository.java
@@ -1,8 +1,8 @@
-package wap.starlist.repository;
+package wap.starlist.bookmark.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import wap.starlist.domain.Root;
+import wap.starlist.bookmark.domain.Root;
 
 @Repository
 public interface RootRepository extends JpaRepository<Root, Long> {
