+++ server/src/main/java/wap/starlist/user/domain/User.java
@@ -1,4 +1,4 @@
-package wap.starlist.domain;
+package wap.starlist.user.domain;
 
 import jakarta.persistence.*;
